refactor: separate likelihood from ancestral reconstruction
tree::asr had grown into general phylogenetic likelihood machinery under a
name describing one consumer of it. Of its four submodules only
reconstruction.rs (27 lines) was about ancestral states; alignment.rs (163
lines) did not reference trees or ancestral states once.

Re-layer along what the code already does:

- alphabet and alignment move to the crate root. Neither had a single
  reference to trees or reconstruction, and models::gtr was reaching back
  into tree::asr::alphabet for its Alphabet, inverting the layering.
- tree::asr -> tree::likelihood. Felsenstein's pruning and the Viterbi
  recursion compute the tree's log-likelihood; reconstruction is the argmax
  taken from it. Rate heterogeneity is likelihood machinery, not an
  ancestral-state feature.
- tree::asr keeps the MarginalAsr and JointAsr traits and re-exports
  Reconstruction, as a thin layer over the engines.

Dependencies now run one way: asr -> likelihood -> models, alphabet,
alignment. The feature gating falls out on the module boundary as a result:
likelihood is concrete in PhyloTree and gates as a block, asr is
trait-level and needs no gate.

No behaviour change; all 99 tests pass unchanged.

Not addressed: the engines still return a Reconstruction, so a caller
wanting only a likelihood must reconstruct states and discard them.
Separating that needs pruning extracted and made generic over RootedTree,
which is a design change rather than a move.
